RC1 release.

+++ script/logDownload.sh
@@ -8,53 +8,60 @@ now_time=$(date +%Y%m%d%H%M%S)
 
 
 if [ -d /nubomed/midpkg/drug-middleware/ ]; then
+   username="nuobo"
    passwd="Nb123456"
    business_logdir="/nubomed/midpkg/logs"
 else
+   username="nubomed"
    passwd="Nb@123456"
-   business_logdir="/nubomed/consumable-service/logs"
+   business_logdir="/nubomed/consumable-cabinet-service/logs"
 fi
 
 if [ -d /nubomed/ecart-service/ ]; then
+   username="nubomed"
    passwd="Nb@123456"
    business_logdir="/nubomed/ecart-service/logs"
 fi
 
 if [ "$4" = "1" ]; then
    for element in $(ls $business_logdir); do
-      if [ "$(stat -c %Y "$business_logdir/$element")" -gt "$start_time" ] && [ "$(stat -c %Y "$business_logdir/$element")" -lt "$end_time" ]; then
+      if [ "$(stat -c %Y "$business_logdir/$element")" -ge "$start_time" ] && [ "$(stat -c %Y "$business_logdir/$element")" -le "$end_time" ]; then
          echo $element
-         zip -r $business_logdir/business_logs$now_time.zip $business_logdir/$element
+         zip -r $business_logdir/business_logs$now_time.zip $business_logdir/$element >>/dev/null
       fi
    done
-   mv $business_logdir/business_logs$now_time.zip $download_dir
+   if [ -f $business_logdir/business_logs$now_time.zip ]; then
+      mv $business_logdir/business_logs$now_time.zip $download_dir
+   else
+      echo "The selected condition range has no logs"
+   fi
 elif [ "$4" = "2" ]; then
+   echo $passwd | sudo -S touch /var/log/sys_logs$now_time.zip
    for element in $(ls /var/log/syslog*); do
-      if [ "$(stat -c %Y "$element")" -gt "$start_time" ] && [ "$(stat -c %Y "$element")" -lt "$end_time" ]; then
-         echo $element
-         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element
+      if [ "$(stat -c %Y "$element")" -ge "$start_time" ] && [ "$(stat -c %Y "$element")" -le "$end_time" ]; then
+         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element >>/dev/null
       fi
    done
-   echo $passwd | sudo -S mv /var/log/sys_logs$now_time.zip $download_dir
-   echo $passwd | sudo -S chown -R nuobo:nuobo $download_dir/sys_logs$now_time.zip
    
    for element in $(ls /var/log/kern*); do
-      if [ "$(stat -c %Y "$element")" -gt "$start_time" ] && [ "$(stat -c %Y "$element")" -lt "$end_time" ]; then
-         echo $element
-         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element
+      if [ "$(stat -c %Y "$element")" -ge "$start_time" ] && [ "$(stat -c %Y "$element")" -le "$end_time" ]; then
+         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element >>/dev/null
       fi
    done
-   echo $passwd | sudo -S mv /var/log/sys_logs$now_time.zip $download_dir
-   echo $passwd | sudo -S chown -R nuobo:nuobo $download_dir/sys_logs$now_time.zip
    
    for element in $(ls /var/log/boot*); do
-      if [ "$(stat -c %Y "$element")" -gt "$start_time" ] && [ "$(stat -c %Y "$element")" -lt "$end_time" ]; then
-         echo $element
-         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element
+      if [ "$(stat -c %Y "$element")" -ge "$start_time" ] && [ "$(stat -c %Y "$element")" -le "$end_time" ]; then
+         echo $passwd | sudo -S zip -r /var/log/sys_logs$now_time.zip $element >>/dev/null
       fi
    done
-   echo $passwd | sudo -S mv /var/log/sys_logs$now_time.zip $download_dir
-   echo $passwd | sudo -S chown -R nuobo:nuobo $download_dir/sys_logs$now_time.zip
+
+   if [ -f /var/log/sys_logs$now_time.zip ]; then
+      echo $passwd | sudo -S mv /var/log/sys_logs$now_time.zip $download_dir
+      echo $passwd | sudo -S chown -R $username:$username $download_dir/sys_logs$now_time.zip >>/dev/null
+   else
+      echo "The selected condition range has no logs"
+   fi
+
 else
    echo "type error"
    exit 1
